refactor: e2e tests for 'Asset Configuration' constraints
